Moved flyskyHallStick_driver.[cpp|h] to common/arm/stm32.
