+++ radio/src/hal/adc_driver.cpp
@@ -136,13 +136,13 @@ bool adcRead()
   uint8_t first_analog_adc;
 
 #if defined(RADIO_FAMILY_T16) || defined(PCBNV14) || defined(RADIO_BOXER)
-  if (globalData.flyskygimbals)
-  {
-      first_analog_adc = FIRST_ANALOG_ADC_FS;
-  } else
-  {
-      first_analog_adc = FIRST_ANALOG_ADC;
-  }
+    if (globalData.flyskygimbals)
+    {
+        first_analog_adc = FIRST_ANALOG_ADC_FS;
+    } else
+    {
+        first_analog_adc = FIRST_ANALOG_ADC;
+    }
 #else
     first_analog_adc = FIRST_ANALOG_ADC;
 #endif
