+++ radio/src/targets/taranis/CMakeLists.txt
@@ -242,7 +242,7 @@ elseif(PCB STREQUAL X7)
     set(PXX2 ON)
     set(ENABLE_SERIAL_PASSTHROUGH ON CACHE BOOL "Enable serial passthrough")
     set(CLI ON CACHE BOOL "Enable CLI")
-    set(FLYSKY_HALL_STICKS_DRIVER ../horus/flyskyHallStick_driver.cpp)
+    set(FLYSKY_HALL_STICKS_DRIVER ../common/arm/stm32/flyskyHallStick_driver.cpp)
   elseif(PCBREV STREQUAL ZORRO)
     set(DEFAULT_INTERNAL_MODULE CROSSFIRE CACHE STRING "Default internal module")
     set(INTERNAL_MODULE_SERIAL YES)
